Moves basicAuth out of lib/utils.js
This is the first step to improving our implementation of middleware and
and testing stragety. It extracts the basicAuth method in utils.js and
places it into its own file. This will be easier to test and manage. At
this point it has only been manually test on a local machine but automated
unit tests will be added next.
