+++ lib/utils.js
@@ -2,7 +2,6 @@
 const fs = require('fs')
 
 // NPM dependencies
-const basicAuth = require('basic-auth')
 const getKeypath = require('keypather/get')
 const marked = require('marked')
 const path = require('path')
@@ -62,37 +61,6 @@ exports.addCheckedFunction = function (env) {
     }
     return checked
   })
-}
-
-/**
- * Simple basic auth middleware for use with Express 4.x.
- *
- * Based on template found at: http://www.danielstjules.com/2014/08/03/basic-auth-with-express-4/
- *
- * @example
- * app.use('/api-requiring-auth', utils.basicAuth('username', 'password'))
- *
- * @param   {string}   username Expected username
- * @param   {string}   password Expected password
- * @returns {function} Express 4 middleware requiring the given credentials
- */
-
-exports.basicAuth = function (username, password) {
-  return function (req, res, next) {
-    if (!username || !password) {
-      console.log('Username or password is not set.')
-      return res.send('<h1>Error:</h1><p>Username or password not set. <a href="https://govuk-prototype-kit.herokuapp.com/docs/publishing-on-heroku#6-set-a-username-and-password">See guidance for setting these</a>.</p>')
-    }
-
-    var user = basicAuth(req)
-
-    if (!user || user.name !== username || user.pass !== password) {
-      res.set('WWW-Authenticate', 'Basic realm=Authorization Required')
-      return res.sendStatus(401)
-    }
-
-    next()
-  }
 }
 
 // Find an available port to run the server on
